tests: split up bash detection library
Split up the detection for whether we're running under bash, and
whether it's the /bin/sh POSIX-y mode or the /bin/bash bash-y mode
into its own library.

This will soon be used very early in test-lib.sh itself to check for
the /bin/bash, so let's make this new file as small as possible.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>